requirements.txt
# 데이터베이스
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.5

# HTTP 클라이언트
requests==2.31.0
aiohttp==3.8.6

# HTML 파싱
beautifulsoup4==4.12.2
lxml==4.9.3

# 유틸리티
python-dotenv==1.0.0
tqdm==4.66.1
retrying==1.3.4

# 로깅 및 모니터링
loguru==0.7.2

# 테스트
pytest==7.4.3
